+++ selection.h
@@ -14,6 +14,7 @@ void selection_start(
 void selection_update(struct terminal *term, int col, int row);
 void selection_finalize(struct terminal *term, uint32_t serial);
 void selection_cancel(struct terminal *term);
+void selection_extend(struct terminal *term, int col, int row, uint32_t serial);
 
 bool selection_on_row_in_view(const struct terminal *term, int row_no);