+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   the Wayland window.
 * **title** option to `footrc`, that sets the initial window title.
 * `--title` command line option, that sets the initial window title.
+* Right mouse button extends the current selection.
 
 ### Changed